Split off Channels, Items and Players
