+++ lib/game/command/say.ex
@@ -7,6 +7,7 @@ defmodule Game.Command.Say do
 
   import Game.Room.Helpers, only: [find_character: 3]
 
+  alias Game.Format.Channels, as: FormatChannels
   alias Game.Hint
   alias Game.Utility
 
@@ -139,7 +140,7 @@ defmodule Game.Command.Say do
 
   def say(state = %{character: character, save: save}, parsed_message) do
     parsed_message = Message.format(parsed_message)
-    state.socket |> @socket.echo(Format.say(:you, parsed_message))
+    state.socket |> @socket.echo(FormatChannels.say(:you, parsed_message))
     save.room_id |> @environment.say({:player, character}, Message.new(character, parsed_message))
   end
 
@@ -158,7 +159,7 @@ defmodule Game.Command.Say do
           |> Map.put(:message, message)
           |> Message.format()
 
-        state.socket |> @socket.echo(Format.say_to(:you, directed_character, parsed_message))
+        state.socket |> @socket.echo(FormatChannels.say_to(:you, directed_character, parsed_message))
 
         room.id
         |> @environment.say({:player, character}, Message.say_to(character, directed_character, parsed_message))
