+++ lib/game/command/shops.ex
@@ -9,6 +9,7 @@ defmodule Game.Command.Shops do
 
   alias Game.Environment
   alias Game.Environment.State.Overworld
+  alias Game.Format.Items, as: FormatItems
   alias Game.Format.Rooms, as: FormatRooms
   alias Game.Items
   alias Game.Utility
@@ -309,7 +310,7 @@ defmodule Game.Command.Shops do
   defp buy_item(shop, item_name, state = %{socket: socket, save: save}) do
     case shop.id |> @shop.buy(item_name, save) do
       {:ok, save, item} ->
-        message = gettext("You bought %{item} from %{shop}.", item: Format.item_name(item), shop: Format.shop_name(shop))
+        message = gettext("You bought %{item} from %{shop}.", item: FormatItems.item_name(item), shop: Format.shop_name(shop))
         socket |> @socket.echo(message)
 
         state = %{state | save: save}
@@ -320,12 +321,12 @@ defmodule Game.Command.Shops do
         socket |> @socket.echo(message)
 
       {:error, :not_enough_currency, item} ->
-        message = gettext("You do not have enough %{currency} for %{item}.", currency: currency(), item: Format.item_name(item))
+        message = gettext("You do not have enough %{currency} for %{item}.", currency: currency(), item: FormatItems.item_name(item))
         socket |> @socket.echo(message)
 
       {:error, :not_enough_quantity, item} ->
         message =
-          gettext("%{shop} does not have enough of %{item} for you to buy.", shop: Format.shop_name(shop), item: Format.item_name(item))
+          gettext("%{shop} does not have enough of %{item} for you to buy.", shop: Format.shop_name(shop), item: FormatItems.item_name(item))
 
         socket |> @socket.echo(message)
     end
@@ -336,7 +337,7 @@ defmodule Game.Command.Shops do
       {:ok, save, item} ->
         message =
           gettext("You sold %{item} to %{shop} for %{cost} %{currency}.", [
-            item: Format.item_name(item),
+            item: FormatItems.item_name(item),
             shop: Format.shop_name(shop),
             cost: item.cost,
             currency: currency()
@@ -362,7 +363,7 @@ defmodule Game.Command.Shops do
         socket |> @socket.echo(message)
 
       item ->
-        socket |> @socket.echo(Format.item(item))
+        socket |> @socket.echo(FormatItems.item(item))
     end
   end
