+++ test/game/format/channels_test.exs
@@ -1,0 +1,7 @@
+defmodule Game.Format.ChannelsTest do
+  use ExUnit.Case
+
+  alias Game.Format.Channels
+
+  doctest Game.Format.Channels
+end
